+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.0] - 2016-02-29
-## Added
+### Added
 - Free draw (#1)
 - Drawing style history (#4)