+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Keeping the clciked position of shape on moveing (#6)
+
 ## [1.1.0] - 2016-02-29
 ### Added
 - Free draw (#1)