+++ CHANGELOG.md
@@ -7,12 +7,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Zoom
 
 ### Changed
-- Keeping the clciked position of shape on moveing (#6)
+- Keeping the clciked position of shape on moveing ([#6](https://github.com/ukatama/nekoboard/issues/6))
+
+### Fixed
+- Fix toolbar sizing
+- Fix scroll on mobile
 
 ## [1.1.0] - 2016-02-29
 ### Added
-- Free draw (#1)
-- Drawing style history (#4)
+- Free draw ([#1](https://github.com/ukatama/nekoboard/issues/1))
+- Drawing style history ([#4](https://github.com/ukatama/nekoboard/issues/4))
 
 ## [1.0.0] - 2016-02-28
 First version.