+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Enhance Color Picker (react-color)
+- Shape which has been changed recently to the top ([#12](https://github.com/ukatama/nekoboard/issues/12)
 
 ## [1.2.0] - 2016-03-02
 ### Added